+++ src/core/diff/index.ts
@@ -1,3 +1,3 @@
 export * from "./DiffPoint";
 export * from "./DiffViewer";
-export * from "./DiffReport";
+export * from "./DiffSummary";